Move file deletion confirmation logic to fileUtils.ts
